Add a new ChefRedundantCode area
Move some of the chefmodernize and chefcorrect cops into this new group. This is for purely redundant code, not code that was made redundant by a newer version of chef. Chef Modernize will still hold cops that require a specific version of Chef before they can be run. This group should be safe regardless of Chef release.
