+++ lib/rubocop/cop/chef/modernize/property_with_name_attribute.rb
@@ -18,8 +18,7 @@ module RuboCop
   module Cop
     module Chef
       module ChefCorrectness
-        # When using properties in a custom resource you should use name_property not
-        # the legacy name_attribute from the days of attributes
+        # When using properties in a custom resource you should use name_property not the legacy name_attribute from the days of attributes
         #
         # @example
         #
