+++ lib/rubocop/cop/chef/redundant/conflicts_metadata.rb
@@ -18,27 +18,25 @@
 module RuboCop
   module Cop
     module Chef
-      module ChefModernize
-        # The long_description metadata.rb method is not used and is unnecessary in cookbooks.
+      module ChefRedundantCode
+        # The conflicts metadata.rb method is not used and is unnecessary in cookbooks.
         #
         # @example
         #
-        #   # bad
-        #   long_description 'this is my cookbook and this description will never be seen'
+        #   # bad in metadata.rb:
         #
-
-        class LongDescriptionMetadata < Cop
-          include RangeHelp
-
-          MSG = 'The long_description metadata.rb method is not used and is unnecessary in cookbooks.'.freeze
+        #   conflicts "another_cookbook"
+        #
+        class ConflictsMetadata < Cop
+          MSG = 'The conflicts metadata.rb method is not used and is unnecessary in cookbooks.'.freeze
 
           def on_send(node)
-            add_offense(node, location: :expression, message: MSG, severity: :refactor) if node.method_name == :long_description
+            add_offense(node, location: :expression, message: MSG, severity: :refactor) if node.method_name == :conflicts
           end
 
           def autocorrect(node)
             lambda do |corrector|
-              corrector.remove(range_with_surrounding_space(range: node.loc.expression, side: :right))
+              corrector.remove(node.loc.expression)
             end
           end
         end
