+++ lib/rubocop/cop/chef/redundant/resource_with_nothing_action.rb
@@ -17,27 +17,26 @@
 module RuboCop
   module Cop
     module Chef
-      module ChefCorrectness
-        # There is no need to define a property named :name in a resource as Chef Infra defines that property for all resources by default.
+      module ChefRedundantCode
+        # Chef Infra Client provides the :nothing action by default for every resource. There is no need to define a :nothing action in your resource code.
         #
         # @example
         #
         #   # bad
-        #   property :name, String
-        #   property :name, String, name_property: true
+        #   action :nothing
+        #     # let's do nothing
+        #   end
         #
-        class UnnecessaryNameProperty < Cop
-          MSG = 'There is no need to define a property named :name in a resource as Chef Infra defines that property for all resources by default.'.freeze
+        class ResourceWithNothingAction < Cop
+          MSG = 'There is no need to define a :nothing action in your resource as Chef Infra Client provides the :nothing action by default for every resource.'.freeze
 
-          def_node_matcher :name_property?, <<-PATTERN
-            (send nil? :property (sym :name) (const nil? :String) $...)
+          def_node_matcher :nothing_action?, <<-PATTERN
+            (block (send nil? :action (sym :nothing)) (args) ... )
           PATTERN
 
-          def on_send(node)
-            name_property?(node) do |hash_vals|
-              if hash_vals.empty? || (hash_vals.first.keys.count == 1 && hash_vals.first.keys.first.source == 'name_property')
-                add_offense(node, location: :expression, message: MSG, severity: :refactor)
-              end
+          def on_block(node)
+            nothing_action?(node) do
+              add_offense(node, location: :expression, message: MSG, severity: :refactor)
             end
           end
