+++ spec/rubocop/cop/chef/redundant/attribute_metadata_spec.rb
@@ -16,7 +16,7 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::Chef::ChefModernize::AttributeMetadata, :config do
+describe RuboCop::Cop::Chef::ChefRedundantCode::AttributeMetadata, :config do
   subject(:cop) { described_class.new(config) }
 
   it 'registers an offense when metadata uses "attribute"' do
