+++ spec/rubocop/cop/chef/redundant/long_description_metadata_spec.rb
@@ -16,16 +16,18 @@
 
 require 'spec_helper'
 
-describe RuboCop::Cop::Chef::ChefModernize::ConflictsMetadata, :config do
+describe RuboCop::Cop::Chef::ChefRedundantCode::LongDescriptionMetadata, :config do
   subject(:cop) { described_class.new(config) }
 
-  it 'registers an offense when metadata uses "conflicts"' do
+  it 'registers an offense when metadata uses "long_description"' do
     expect_offense(<<~RUBY)
-      conflicts 'foo'
-      ^^^^^^^^^^^^^^^ The conflicts metadata.rb method is not used and is unnecessary in cookbooks.
+      description 'foo'
+      long_description 'foo'
+      ^^^^^^^^^^^^^^^^^^^^^^ The long_description metadata.rb method is not used and is unnecessary in cookbooks.
+      version '1.0.0'
     RUBY
 
-    expect_correction("\n")
+    expect_correction("description 'foo'\nversion '1.0.0'\n")
   end
 
   it "doesn't register an offense on normal metadata" do
